Slight resturcutre of extraction pipeline - adding reprocessing_utils.py
